refactor(api): split model into files
