+++ api/src/rpc/model/admin.rs
@@ -1,0 +1,1 @@
+
